Refactor resync path code out of PGSecondary class
